updated cmus

+++ .zprezto/modules/prompt/external/powerlevel10k/internal/wizard.zsh
@@ -281,6 +281,7 @@ function quit() {
 }
 
 local screen_widgets=()
+local -i max_priority
 local -i prompt_idx
 local choice
 
@@ -289,6 +290,7 @@ function add_widget() {
   shift
   local render="${(j: :)${(@q)*}}"
   screen_widgets+=("$priority" "$render")
+  (( priority <= max_priority )) || max_priority=priority
 }
 
 function render_screen_pass() {
@@ -327,7 +329,7 @@ function render_screen() {
         else
           break
         fi
-        while (( (COLUMNS > 88 ? 88 : COLUMNS) == wizard_columns && LINES == wizard_lines )); do
+        while (( get_columns() == wizard_columns && LINES == wizard_lines )); do
           sleep 1
         done
       done
@@ -356,7 +358,7 @@ function render_screen() {
       flowing -c %1FNot enough vertical space.%f
       print
       flowing Make terminal window %Btaller%b or press %BCtrl-C%b to abort Powerlevel10k configuration wizard.
-      while (( (COLUMNS > 88 ? 88 : COLUMNS) == wizard_columns && LINES == wizard_lines )); do
+      while (( get_columns() == wizard_columns && LINES == wizard_lines )); do
         sleep 1
       done
     done
@@ -389,6 +391,7 @@ function ask() {
   local -i lines columns wizard_lines wizard_columns
   add_widget 0 print -P "(q)  Quit and do nothing."
   add_widget 0 print
+  add_widget $((max_priority + 1))
   add_widget 0 print -P "%BChoice [${choices}q]: %b"
   while true; do
     =true
@@ -404,6 +407,7 @@ function ask() {
       fi
       if [[ $choices == *$choice* ]]; then
         screen_widgets=()
+        max_priority=0
         prompt_idx=0
         return
       fi
@@ -501,41 +505,42 @@ function install_font() {
   clear
   case $terminal in
     Termux)
-      mkdir -p ~/.termux || quit -c
+      command mkdir -p -- ~/.termux || quit -c
       run_command "Downloading %BMesloLGS NF Regular.ttf%b" \
         curl -fsSL -o ~/.termux/font.ttf "$font_base_url/MesloLGS%20NF%20Regular.ttf"
       run_command "Reloading %BTermux%b settings" termux-reload-settings
     ;;
     iTerm2)
-      mkdir -p ~/Library/Fonts || quit -c
+      command mkdir -p -- ~/Library/Fonts || quit -c
       local style
       for style in Regular Bold Italic 'Bold Italic'; do
         local file="MesloLGS NF ${style}.ttf"
         run_command "Downloading %B$file%b" \
           curl -fsSL -o ~/Library/Fonts/$file.tmp "$font_base_url/${file// /%20}"
-        zf_mv -f -- ~/Library/Fonts/$file{.tmp,} || quit -c
+        command mv -f -- ~/Library/Fonts/$file{.tmp,} || quit -c
       done
       print -nP -- "Changing %BiTerm2%b settings ..."
       local size=$iterm2_font_size
       [[ $size == 12 ]] && size=13
       local k t v settings=(
-        '"Normal Font"'            string '"MesloLGS-NF-Regular '$size'"'
-        '"Terminal Type"'          string '"xterm-256color"'
-        '"Horizontal Spacing"'     real   1
-        '"Vertical Spacing"'       real   1
-        '"Minimum Contrast"'       real   0
-        '"Use Bold Font"'          bool   1
-        '"Use Bright Bold"'        bool   1
-        '"Use Italic Font"'        bool   1
-        '"ASCII Anti Aliased"'     bool   1
-        '"Non-ASCII Anti Aliased"' bool   1
-        '"Use Non-ASCII Font"'     bool   0
-        '"Ambiguous Double Width"' bool   0
-        '"Draw Powerline Glyphs"'  bool   1
+        '"Normal Font"'                                 string '"MesloLGS-NF-Regular '$size'"'
+        '"Terminal Type"'                               string '"xterm-256color"'
+        '"Horizontal Spacing"'                          real   1
+        '"Vertical Spacing"'                            real   1
+        '"Minimum Contrast"'                            real   0
+        '"Use Bold Font"'                               bool   1
+        '"Use Bright Bold"'                             bool   1
+        '"Use Italic Font"'                             bool   1
+        '"ASCII Anti Aliased"'                          bool   1
+        '"Non-ASCII Anti Aliased"'                      bool   1
+        '"Use Non-ASCII Font"'                          bool   0
+        '"Ambiguous Double Width"'                      bool   0
+        '"Draw Powerline Glyphs"'                       bool   1
+        '"Only The Default BG Color Uses Transparency"' bool   1
       )
       for k t v in $settings; do
         /usr/libexec/PlistBuddy -c "Set :\"New Bookmarks\":0:$k $v" \
-          ~/Library/Preferences/com.googlecode.iterm2.plist && continue
+          ~/Library/Preferences/com.googlecode.iterm2.plist 2>/dev/null && continue
         run_command "" /usr/libexec/PlistBuddy -c \
           "Add :\"New Bookmarks\":0:$k $t $v" ~/Library/Preferences/com.googlecode.iterm2.plist
       done
@@ -665,13 +670,13 @@ function ask_diamond() {
   add_widget 0 print
   add_widget 0 flowing -c -- "--->  \uE0B2\uE0B0  <---"
   add_widget 0 print
-  add_widget 1
+  add_widget 3
   add_widget 0 print -P "%B(y)  Yes.%b"
   add_widget 0 print
   add_widget 1
   add_widget 0 print -P "%B(n)  No.%b"
   add_widget 0 print
-  add_widget 1
+  add_widget 2
   if (( can_install_font )); then
     extra+=r
     add_widget 0 print -P "(r)  Restart from the beginning."
@@ -693,10 +698,13 @@ function ask_lock() {
   add_widget 0 print
   add_widget 0 flowing -c -- "--->  $1  <---"
   add_widget 0 print
+  add_widget 3
   add_widget 0 print -P "%B(y)  Yes.%b"
   add_widget 0 print
+  add_widget 1
   add_widget 0 print -P "%B(n)  No.%b"
   add_widget 0 print
+  add_widget 2
   add_widget 0 print -P "(r)  Restart from the beginning."
   ask ynr
   case $choice in
@@ -713,10 +721,13 @@ function ask_python() {
   add_widget 0 print -P ""
   add_widget 0 flowing -c -- "--->  \uE63C  <---"
   add_widget 0 print -P ""
+  add_widget 3
   add_widget 0 print -P "%B(y)  Yes.%b"
   add_widget 0 print -P ""
+  add_widget 1
   add_widget 0 print -P "%B(n)  No.%b"
   add_widget 0 print -P ""
+  add_widget 2
   add_widget 0 print -P "(r)  Restart from the beginning."
   ask ynr
   case $choice in
@@ -732,10 +743,13 @@ function ask_arrow() {
   add_widget 0 print -P ""
   add_widget 0 flowing -c -- "--->  \u276F\u276E  <---"
   add_widget 0 print -P ""
+  add_widget 3
   add_widget 0 print -P "%B(y)  Yes.%b"
   add_widget 0 print -P ""
+  add_widget 1
   add_widget 0 print -P "%B(n)  No.%b"
   add_widget 0 print -P ""
+  add_widget 2
   add_widget 0 print -P "(r)  Restart from the beginning."
   ask ynr
   case $choice in
@@ -752,10 +766,13 @@ function ask_debian() {
   add_widget 0 print -P ""
   add_widget 0 flowing -c -- "--->  \uF306  <---"
   add_widget 0 print -P ""
+  add_widget 3
   add_widget 0 print -P "%B(y)  Yes.%b"
   add_widget 0 print -P ""
+  add_widget 1
   add_widget 0 print -P "%B(n)  No.%b"
   add_widget 0 print -P ""
+  add_widget 2
   add_widget 0 print -P "(r)  Restart from the beginning."
   ask ynr
   case $choice in
@@ -786,10 +803,13 @@ function ask_icon_padding() {
   add_widget 0 print -P ""
   add_widget 0 flowing -c -- "--->  $text  <---"
   add_widget 0 print -P ""
+  add_widget 3
   add_widget 0 flowing +c -i 5 "%B(y)  Yes." Icons are very close to the crosses but there is "%b%2Fno overlap%f%B.%b"
   add_widget 0 print -P ""
+  add_widget 1
   add_widget 0 flowing +c -i 5 "%B(n)  No." Some icons "%b%2Foverlap%f%B" neighbouring crosses.%b
   add_widget 0 print -P ""
+  add_widget 2
   add_widget 0 print -P "(r)  Restart from the beginning."
   ask ynr
   case $choice in
@@ -1046,6 +1066,8 @@ function os_icon_name() {
           local lines=(${(f)"$(</etc/os-release)"})
           lines=(${(@M)lines:#ID=*})
           (( $#lines == 1 )) && os_release_id=${lines[1]#ID=}
+        elif [[ -e /etc/artix-release ]]; then
+          os_release_id=artix
         fi
         case $os_release_id in
           *arch*)                  echo LINUX_ARCH_ICON;;
@@ -1068,6 +1090,7 @@ function os_icon_name() {
           *devuan*)                echo LINUX_DEVUAN_ICON;;
           *manjaro*)               echo LINUX_MANJARO_ICON;;
           *void*)                  echo LINUX_VOID_ICON;;
+          *artix*)                 echo LINUX_ARTIX_ICON;;
           *)                       echo LINUX_ICON;;
         esac
         ;;
@@ -1388,16 +1411,16 @@ function ask_empty_line() {
   add_widget 0 print -P "%B(1)  Compact.%b"
   add_widget 0 print
   add_widget 1
-  add_widget 0 print_prompt
-  add_widget 0 print_prompt
+  add_prompt_n
+  add_prompt_n
   add_widget 0 print
   add_widget 2
   add_widget 0 print -P "%B(2)  Sparse.%b"
   add_widget 0 print
   add_widget 1
-  add_widget 0 print_prompt
+  add_prompt_n
   add_widget 0 print
-  add_widget 0 print_prompt
+  add_prompt_n
   add_widget 0 print
   add_widget 2
   add_widget 0 print -P "(r)  Restart from the beginning."
@@ -1420,34 +1443,37 @@ function print_instant_prompt_link() {
 function ask_instant_prompt() {
   if ! is-at-least 5.4; then
     instant_prompt=off
+    options+=instant_prompt=auto-off
     return 0
+  fi
+  if (( $+functions[z4h] )); then
+    instant_prompt=quiet
+    options+=instant_prompt=auto-quiet
+    return
   fi
   add_widget 0 flowing -c "%BInstant Prompt Mode%b"
   add_widget 0 print_instant_prompt_link
   add_widget 1
   add_widget 0 print
   add_widget 2
-  add_widget 0 flowing +c -i 5 "%B(1)  Off.%b" Disable instant prompt. Choose this if you\'ve \
-    tried instant prompt and found it incompatible with your zsh configuration files.
+  add_widget 0 flowing +c -i 5 "%B(1)  Verbose (recommended).%b"
   add_widget 0 print
   add_widget 1
-  add_widget 0 flowing +c -i 5 "%B(2)  Quiet.%b" Enable instant prompt and %Bdon\'t print      \
-    warnings%b when detecting console output during zsh initialization. Choose this if you\'ve \
-    read and understood instant prompt documentation.
+  add_widget 0 flowing +c -i 5 "%B(2)  Quiet.%b" Choose this if you\'ve read and understood \
+    instant prompt documentation.
   add_widget 0 print
   add_widget 1
-  add_widget 0 flowing +c -i 5 "%B(3)  Verbose.%b"  Enable instant prompt and %Bprint a warning%b \
-    when detecting console output during zsh initialization. %BChoose this if you\'ve never tried \
-    instant prompt, haven\'t seen the warning, or if you are unsure what this all means%b.
+  add_widget 0 flowing +c -i 5 "%B(3)  Off.%b" Choose this if you\'ve tried instant prompt \
+    and found it incompatible with your zsh configuration files.
   add_widget 0 print
   add_widget 2
   add_widget 0 print -P "(r)  Restart from the beginning."
   ask 123r
   case $choice in
     r) return 1;;
-    1) instant_prompt=off; options+=instant_prompt=off;;
+    1) instant_prompt=verbose; options+=instant_prompt=verbose;;
     2) instant_prompt=quiet; options+=instant_prompt=quiet;;
-    3) instant_prompt=verbose; options+=instant_prompt=verbose;;
+    3) instant_prompt=off; options+=instant_prompt=off;;
   esac
   return 0
 }
@@ -1564,7 +1590,11 @@ function ask_zshrc_edit() {
   add_widget 0 print -P ""
   add_widget 1
   local modifiable=y
-  if [[ -e $__p9k_zshrc && ! -w $__p9k_zshrc ]]; then
+  if [[ ! -w $__p9k_zd ]]; then
+    modifiable=
+    add_widget 0 flowing -c %3FWARNING:%f %2F${__p9k_zd_u//\\/\\\\}%f %3Fis readonly.%f
+    add_widget 0 print -P ""
+  elif [[ -e $__p9k_zshrc && ! -w $__p9k_zshrc ]]; then
     local -a stat
     zstat -A stat +uid -- $__p9k_zshrc || quit -c
     if (( stat[1] == EUID )); then
@@ -1849,6 +1879,8 @@ function generate_config() {
   header+=$line
   header+=$'.\n# Type `p10k configure` to generate another config.\n#'
 
+  command mkdir -p -- ${__p9k_cfg_path:h} || return
+
   if [[ -e $__p9k_cfg_path ]]; then
     unlink $__p9k_cfg_path || return
   fi
@@ -1888,7 +1920,7 @@ fi" || return
 [[ ! -f ${(%)__p9k_cfg_path_u} ]] || source ${(%)__p9k_cfg_path_u}" || return
     fi
     (( writable )) || chmod u-w -- $tmp || return
-    zf_mv -f -- $tmp $__p9k_zshrc || return
+    command mv -f -- $tmp $__p9k_zshrc || return
   } always {
     zf_rm -f -- $tmp
   }
